Make wgrep-ag work when ag-group-matches is on
ag-group-matches being the default in newer ag.el, as far as I can
tell.

This uses what might well be considered internals of compile.el, but
by doing so we basically put the onus of providing data about matches
on ag.el, and then we just read what it has provided to
compilation-mode.  This *should* work with both grouped and ungrouped
results.

Aims to fix mhayashi1120/Emacs-wgrep#30.